Bytt namn på modellen Image till ProductImage. Ny metod i DatabaseManager för bilder.
